Rename generic.{hpp,cpp} to misc.{hpp,cpp}
