+++ include/chemfiles/misc.hpp
